+++ contracts/staking/src/state.rs
@@ -42,7 +42,7 @@ pub struct InvestmentInfo {
     /// this is how much the owner takes as a cut when someone unbonds
     pub exit_tax: Decimal,
     /// All tokens are bonded to this validator
-    /// FIXME: humanize/canonicalize address doesn't work for validator addrresses
+    /// addr_humanize/addr_canonicalize doesn't work for validator addrresses (e.g. cosmosvaloper1...)
     pub validator: String,
     /// This is the minimum amount we will pull out to reinvest, as well as a minumum
     /// that can be unbonded (to avoid needless staking tx)